keyboard interrupts almost working; need to convert scancodes
Okay, so this was a major commit! I finally figured out how to implement a (basic) keyboard driver! I have taken a look at the wiki, and it is a wealth of good information for working towards the best and most robust keyboard driver, so I will work on that. However, before I can attempt anything like that, I must first add support for scancodes, because currently I just return the raw scancodes and don't convert them to characters. After that, I will work on the project structure yet again, and fix the age-old issues with the console module. (i will probably move  into the  submodule because it makes sense).